One lake testing
This code does not fix anything for Trino. This was simply a test to check how incompatible OneLake is with how we have azure sdk implemented in trino-filesystem-azure. the test was simply done with a create-schema sql and it ended up breaking at multiple places over the course of the code.

the AzureOneLake standalone java client was a simple test to confirm that service principal based auth indeed works for OneLake